RavenDB-18979 Add disk stats getter for Windows.
