+++ src/Sparrow.Server/Utils/DiskStatsGetter/IDiskStatsGetter.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Sparrow.Server.Utils.DiskStatsGetter;
+
+public interface IDiskStatsGetter
+{
+    DiskStatsResult Get(string drive);
+    Task<DiskStatsResult> GetAsync(string drive);
+}
