+++ src/Sparrow.Server/Utils/DiskStatsGetter/IDiskStatsRawResult.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Sparrow.Server.Utils.DiskStatsGetter;
+
+internal interface IDiskStatsRawResult
+{
+    DateTime Time { get; }
+}
