+++ src/Sparrow.Server/Utils/DiskStatsGetter/NotImplementedDiskStatsGetter.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace Sparrow.Server.Utils.DiskStatsGetter;
+
+internal class NotImplementedDiskStatsGetter : IDiskStatsGetter
+{
+    public DiskStatsResult Get(string drive) => null;
+    public Task<DiskStatsResult> GetAsync(string drive) => null;
+}
