+++ test/SlowTests/Monitoring/DiskStatsGetterTest.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using FastTests;
-using Sparrow.Server.Utils;
+using Sparrow.Server.Utils.DiskStatsGetter;
 using Tests.Infrastructure;
 using Xunit;
 using Xunit.Abstractions;
@@ -19,11 +19,23 @@ namespace SlowTests.Monitoring
         }
 
         [NightlyBuildMultiplatformFact(RavenPlatform.Linux)]
-        public async Task DiskStats_WhenGet_ShouldBeLessThenTwoSimpleGet()
+        public async Task LinuxDiskStats_WhenGet_ShouldBeLessThenTwoSimpleGet()
+        {
+            var diskStatsGetter = new LinuxDiskStatsGetter(TimeSpan.FromMilliseconds(100));
+            await DiskStats_WhenGet_ShouldBeLessThenTwoSimpleGet(diskStatsGetter);
+        }
+        
+        [NightlyBuildMultiplatformFact(RavenPlatform.Windows)]
+        public async Task WindowsDiskStats_WhenGet_ShouldBeLessThenTwoSimpleGet()
+        {
+            var diskStatsGetter = new WindowsDiskStatsGetter(TimeSpan.FromMilliseconds(100));
+            await DiskStats_WhenGet_ShouldBeLessThenTwoSimpleGet(diskStatsGetter);
+        }
+
+        private static async Task DiskStats_WhenGet_ShouldBeLessThenTwoSimpleGet(IDiskStatsGetter diskStatsGetter)
         {
             var currentDirectory = Directory.GetCurrentDirectory();
-            var diskStatsGetter = new DiskStatsGetter(TimeSpan.FromMilliseconds(100));
-
+            
             var baseTime = TimeSpan.Zero;
             const int controlGroupCount = 10;
             for (int i = 0; i < controlGroupCount; i++)
